Get latest access token for the listing API
The access token is required to call the domain.com.au listing API. It expires
after 12 hours. The credentials need to be set up by signing up for a developer
account on https://developer.domain.com.au/docs.

It expires every 12 hours so I automated the renewal process and have my
credentials stored in a separate file (.gitignore) so that it doesn't get
uploaded to GitHub.